+++ notdo-domain/src/main/kotlin/com/gsm/notdo/domain/not_todo/model/History.kt
@@ -1,8 +1,10 @@
 package com.gsm.notdo.domain.not_todo.model
 
+import com.gsm.notdo.common.annotation.Aggregate
 import java.time.LocalDateTime
 import java.util.*
 
+@Aggregate
 data class History(
         val id: UUID,
         val isComplete: String,
