--- notdo-infrastructure/src/main/kotlin/com/gsm/notdo/persistence/auth/repository/AuthenticationRepository.kt
@@ -1,7 +1,0 @@
-package com.gsm.notdo.persistence.auth.repository
-
-import com.gsm.notdo.persistence.auth.entity.AuthenticationEntity
-import org.springframework.data.repository.CrudRepository
-
-interface AuthenticationRepository : CrudRepository<AuthenticationEntity, String> {
-}
