--- notdo-infrastructure/src/main/kotlin/com/gsm/notdo/persistence/auth/repository/RefreshTokenRepository.kt
@@ -1,7 +1,0 @@
-package com.gsm.notdo.persistence.auth.repository
-
-import com.gsm.notdo.persistence.auth.entity.RefreshTokenEntity
-import org.springframework.data.repository.CrudRepository
-
-interface RefreshTokenRepository : CrudRepository<RefreshTokenEntity, String> {
-}
