--- notdo-infrastructure/src/main/kotlin/com/gsm/notdo/persistence/category/repository/CategoryRepository.kt
@@ -1,7 +1,0 @@
-package com.gsm.notdo.persistence.category.repository
-
-import com.gsm.notdo.persistence.category.entity.CategoryEntity
-import org.springframework.data.repository.CrudRepository
-
-interface CategoryRepository : CrudRepository<CategoryEntity, Long> {
-}
