--- notdo-infrastructure/src/main/kotlin/com/gsm/notdo/persistence/common/BaseUUIDEntity.kt
@@ -1,17 +1,0 @@
-package com.gsm.notdo.persistence.common
-
-import org.hibernate.annotations.GenericGenerator
-import java.util.UUID
-import javax.persistence.Column
-import javax.persistence.GeneratedValue
-import javax.persistence.Id
-import javax.persistence.MappedSuperclass
-
-@MappedSuperclass
-open class BaseUUIDEntity(
-        @Id
-        @GeneratedValue(generator = "uuid2")
-        @GenericGenerator(name = "uuid2", strategy = "uuid2")
-        @Column(columnDefinition = "BINARY(16)", nullable = false)
-        open val id: UUID
-)
