--- notdo-infrastructure/src/main/kotlin/com/gsm/notdo/persistence/diary/repository/DiaryRepository.kt
@@ -1,8 +1,0 @@
-package com.gsm.notdo.persistence.diary.repository
-
-import com.gsm.notdo.persistence.diary.entity.DiaryEntity
-import org.springframework.data.repository.CrudRepository
-import java.util.UUID
-
-interface DiaryRepository : CrudRepository<DiaryEntity, UUID> {
-}
