+++ notdo-presentation/src/main/kotlin/com/gsm/notdo/common/base/entity/BaseEntity.kt
@@ -1,0 +1,13 @@
+package com.gsm.notdo.common.base.entity
+
+import java.time.LocalDate
+import java.util.*
+import javax.persistence.Column
+import javax.persistence.MappedSuperclass
+
+@MappedSuperclass
+open class BaseEntity (
+        id: UUID,
+        @Column(nullable = false, updatable = false, columnDefinition = "DATETIME(6)")
+        open val createdAt: LocalDate = LocalDate.now()
+) : BaseUUIDEntity(id)
