+++ notdo-presentation/src/main/kotlin/com/gsm/notdo/common/base/entity/BaseUUIDEntity.kt
@@ -1,0 +1,17 @@
+package com.gsm.notdo.common.base.entity
+
+import org.hibernate.annotations.GenericGenerator
+import java.util.UUID
+import javax.persistence.Column
+import javax.persistence.GeneratedValue
+import javax.persistence.Id
+import javax.persistence.MappedSuperclass
+
+@MappedSuperclass
+open class BaseUUIDEntity(
+        @Id
+        @GeneratedValue(generator = "uuid2")
+        @GenericGenerator(name = "uuid2", strategy = "uuid2")
+        @Column(columnDefinition = "BINARY(16)", nullable = false)
+        open val id: UUID
+)
