+++ notdo-presentation/src/main/kotlin/com/gsm/notdo/domain/auth/output/persistence/auth/repository/AuthCodeRepository.kt
@@ -1,0 +1,7 @@
+package com.gsm.notdo.domain.user.output.persistence.auth.repository
+
+import com.gsm.notdo.domain.auth.output.persistence.auth.entity.AuthCodeEntity
+import org.springframework.data.repository.CrudRepository
+
+interface AuthCodeRepository : CrudRepository<AuthCodeEntity, String> {
+}
