+++ notdo-presentation/src/main/kotlin/com/gsm/notdo/domain/auth/output/persistence/auth/repository/RefreshTokenRepository.kt
@@ -1,0 +1,7 @@
+package com.gsm.notdo.domain.user.output.persistence.auth.repository
+
+import com.gsm.notdo.domain.auth.output.persistence.auth.entity.RefreshTokenEntity
+import org.springframework.data.repository.CrudRepository
+
+interface RefreshTokenRepository : CrudRepository<RefreshTokenEntity, String> {
+}
