+++ notdo-presentation/src/main/kotlin/com/gsm/notdo/domain/user/output/persistence/user/repository/UserRepository.kt
@@ -1,0 +1,9 @@
+package com.gsm.notdo.domain.user.output.persistence.user.repository
+
+import com.gsm.notdo.domain.user.output.persistence.user.entity.UserEntity
+import org.springframework.data.repository.CrudRepository
+import java.util.*
+
+interface UserRepository : CrudRepository<UserEntity, UUID> {
+    fun findByUserId(userId: UUID): UserEntity?
+}
