+++ settings.gradle.kts
@@ -1,6 +1,6 @@
 rootProject.name = "notdo"
 
-    include("notdo-domain")
+include("notdo-domain")
 include("notdo-application")
 include("notdo-infrastructure")
 include("notdo-presentation")
